+++ packages/react-accordion/src/components/AccordionItem/useAccordionItemStyles.ts
@@ -1,0 +1,10 @@
+import { mergeClasses } from '@fluentui/react-make-styles';
+import type { AccordionItemState } from './AccordionItem.types';
+
+export const accordionItemClassName = 'fui-AccordionItem';
+
+export const useAccordionItemStyles = (state: AccordionItemState) => {
+  state.root.className = mergeClasses(accordionItemClassName, state.root.className);
+
+  return state;
+};